Update imports for galaxy.util.path moves
