+++ lib/galaxy/objectstore/s3.py
@@ -22,11 +22,11 @@ except ImportError:
 from galaxy.exceptions import ObjectInvalid, ObjectNotFound
 from galaxy.util import (
     directory_hash_id,
-    safe_relpath,
     string_as_bool,
     umask_fix_perms,
     which,
 )
+from galaxy.util.path import safe_relpath
 from galaxy.util.sleeper import Sleeper
 
 from .s3_multipart_upload import multipart_upload
